canvas expansion rework phase 1 complete

+++ crates/darkly/tests/paint_ops.rs
@@ -269,7 +269,7 @@ fn gpu_flood_fill_interior() {
         &mut enc,
         &pipelines,
         &queue,
-        [0, 0, w, h],
+        [0, 0, w as i32, h as i32],
         [0, 0, 255, 255],
         &mask_bg,
     );
@@ -376,7 +376,7 @@ fn gpu_flood_fill_undo() {
         &mut enc,
         &pipelines,
         &queue,
-        [0, 0, w, h],
+        [0, 0, w as i32, h as i32],
         [0, 255, 0, 255],
         &mask_bg,
     );
@@ -603,7 +603,7 @@ fn gpu_fill_rect_with_mask() {
         &mut enc,
         &pipelines,
         &queue,
-        [0, 0, w, h],
+        [0, 0, w as i32, h as i32],
         [0, 255, 0, 255],
         &mask_bg,
     );
